File out Windows HTTP Server package in new format with class constants
The 'Windows HTTP Server' contains a large number of structure classes that
define many class constants representing the offsets of the fields of the
structures. Under the pre-7.2 system of installing the constants after
defining the class as part of separate individual chunks, this package was
rather slow to load, generating many method recompilation scans.