Revert PluggableEnvironment change that had large perf impact
In order to fix the lack of progress updates when browsing the methods
in the '*-compilation issues' category, I made a small change to the way
PluggableEnvironment enumerates classes so that it went through the
BrowserEnvironmentWrapper>>allClassesDo: method, and so reported progress.
However this has a massive performance impact - more than an order of
magnitude slow down showing in PluggableEnvironmentTest>>#testReferencesTo.