Inline `hash:max:` into `findKeyOrNil:` and `findElementOrNil:` methods
If a hashed collection subclass of Set needs to override hash:max: in
order to use a different comparison then it will need to override either
findElementOrNil: (sets) or findKeyOrNil: (dictionaries) anyway, because
the equality comparison has to change with the hash. Therefore there is no
point in paying the cost of an extra message send in this very perf
critical part of the system.