Use clean version of at:ifAbsentPut: by preference
Dictionary>>at:ifAbsentPut: has to look up the key twice because it
assumes that the block passed as the 2nd argument may modify the
collection. Of course it usually doesn't, so this can be avoided the vast
majority of the time.
Need to come up with a better selector than #at:ifAbsentPutClean: though.