Incorrect assert in ANSI tests that 2 objects identityHash are never equal
Identity hashes are typically allocated randomly on deman from a fairly
small range. In Dolphin only 16-bits are available for identity hash
storage, so it is to be expected that several objects in the image will
have the same identity hash. Therefore the assertion that a newly allocated
Object must never have the same identity hash as some pre-existing test
object is incorrect. I'm surprised it hasn't failed before, or has failed
so rarely as to go unnoticed.