Partition base system into Namespaces: Part 62b, IDE
Mainly not namespaced yet, but changes required from introduction
of namespaces in other code, but most RB extension classes are
put in Refactory.Browser.

The newish Smalltalk Parser tests package is renamed and compiler
tests added to it. These tests are mainly of Kernel content, not IDE,
and so are in Kernel.Tests, and moved to System/Compiler/Tests folder.

Not bootable.